Increased flexibility of data included in file, better integration of URM data, replaced double->float throughout, preallocated variables/arrays used in calculating team scores to try to increase optimization speed, cosmetic upgrades

+++ GA.h
@@ -20,7 +20,7 @@ const int mutationLikelihood = 50 * (RAND_MAX/100);     // first number gives pr
 struct tourneyPlayer
 {
     int genome[maxRecords];
-    double score;
+    float score;
 };
